write our own ModuleExec.
The main reason to start down this path is so we can add all the timing
decorators around the exec's.  But also it caused some of the error
handling to shift, and now at least one of our log messages is better.

But this is probably where I need to put things down for the night.
We already have that one TODO about statements vs granularity of the
args, and that issue is pretty fundamental and requires some decisions.
Like, are we going to report times *per thing exec'd*, always?
Or per statement (different)?  Or per *arg* (yet again different)?
What we started with here is the first of those three options, and
probably the least interesting (highest noise per bit of signal).

Signed-off-by: Eric Myhre <hash@exultant.us>