cmd/smsh: don't use parser.Stmts
We want to parse one argument at a time, so parser.Parse is more
appropriate.

The only caveat is that it parses many statements, not just a single
one. However, the old code behaved the same way, so there's no reason
not to do this change.

Another advantage of this mechanism is that we can give a filename, such
that syntax nodes have positions that tell the user which argument they
belong to. This is left for another commit, though.