Absorb some upstream library changes.
(These are all tips of the repos as fetched on March 22.
Things lay on my local working copy for a while uncommitted!
I'll do another yank momentarily that gets us to the present.)

A few minor API changes in libraries, but as far as the quick tests
we have go, nothing seems massively changed.

Signed-off-by: Eric Myhre <hash@exultant.us>